refactor(api): move create method to prescription/target-profile
