+++ api/src/prescription/target-profile/infrastructure/repositories/target-profile-administration-repository.js
@@ -1,8 +1,11 @@
 import _ from 'lodash';
 
 import { knex } from '../../../../../db/knex-database-connection.js';
+import { DomainTransaction } from '../../../../shared/domain/DomainTransaction.js';
 import { NotFoundError, ObjectValidationError } from '../../../../shared/domain/errors.js';
 import { TargetProfile } from '../../../../shared/domain/models/index.js';
+
+const TARGET_PROFILE_TABLE = 'target-profiles';
 
 const update = async function (targetProfile) {
   let results;
@@ -24,4 +27,28 @@ const update = async function (targetProfile) {
   return new TargetProfile(results[0]);
 };
 
-export { update };
+const create = async function ({ targetProfileForCreation }) {
+  const knexConn = DomainTransaction.getConnection();
+  const targetProfileRawData = _.pick(targetProfileForCreation, [
+    'name',
+    'category',
+    'description',
+    'comment',
+    'isPublic',
+    'imageUrl',
+    'ownerOrganizationId',
+    'areKnowledgeElementsResettable',
+  ]);
+  const [{ id: targetProfileId }] = await knexConn(TARGET_PROFILE_TABLE).insert(targetProfileRawData).returning('id');
+
+  const tubesData = targetProfileForCreation.tubes.map((tube) => ({
+    targetProfileId,
+    tubeId: tube.id,
+    level: tube.level,
+  }));
+  await knexConn.batchInsert('target-profile_tubes', tubesData);
+
+  return targetProfileId;
+};
+
+export { create, update };
